feat(gasket-intl): rename IntlProvider to IntlManager
