+++ packages/gasket-intl/test/intl-manager.browser.test.js
@@ -12,7 +12,7 @@ const expectedRegister = {
   'locales/ar-AE': { key: 'قيمة' }
 };
 
-describe('IntlManager', () => {
+describe('InternalIntlManager', () => {
   let IntlManager, initSpy, loadSpy, getElementByIdSpy;
   let manager, mockManifest;
 
@@ -27,8 +27,8 @@ describe('IntlManager', () => {
       }
     };
 
-    const mod = await import('../lib/intl-manager.js');
-    IntlManager = mod.IntlManager;
+    const mod = await import('../lib/internal-intl-manager.js');
+    IntlManager = mod.InternalIntlManager;
     initSpy = jest.spyOn(IntlManager.prototype, 'init');
     loadSpy = jest.spyOn(IntlManager.prototype, 'load');
